+++ middlewares/metrics.js
@@ -1,0 +1,36 @@
+import client from 'prom-client';
+
+// Create a Registry which registers the metrics
+const register = new client.Registry();
+
+// Add a default label which is added to all metrics
+register.setDefaultLabels({
+  app: 'account-backend'
+});
+
+// Enable the collection of default metrics
+client.collectDefaultMetrics({ register });
+
+// Create a custom histogram metric
+const httpRequestDurationMicroseconds = new client.Histogram({
+  name: 'http_request_duration_seconds',
+  help: 'Duration of HTTP requests in seconds',
+  labelNames: ['method', 'route', 'code'],
+  buckets: [0.1, 0.3, 0.5, 0.7, 1, 3, 5, 7, 10]
+});
+
+// Register the histogram
+register.registerMetric(httpRequestDurationMicroseconds);
+
+export const metricsMiddleware = (req, res, next) => {
+  const end = httpRequestDurationMicroseconds.startTimer();
+  res.on('finish', () => {
+    end({ method: req.method, route: req.route ? req.route.path : req.path, code: res.statusCode });
+  });
+  next();
+};
+
+export const getMetrics = async (req, res) => {
+  res.set('Content-Type', register.contentType);
+  res.end(await register.metrics());
+};